+++ index.js
@@ -211,7 +211,7 @@ var IonIcon = /** @class */ (function (_super) {
         name: PropTypes.string.isRequired,
         color: PropTypes.string,
         size: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-        iconType: PropTypes.oneOf(['ios', 'md']),
+        mode: PropTypes.oneOf(['ios', 'md']),
         innerRef: PropTypes.func,
     };
     return IonIcon;